Support for exporting commands with white spaces
export var="ls        -l" will now export correctly.
In the env, is included with white spaces.
if input $var as command in minishell, will treat it as if it was "ls -l".
Still need to add support for echo $var and echo "$var" as one should appear with spaces and other one without.